requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sentence-transformers>=2.7.0
pdfplumber>=0.11.0
numpy>=1.26.0
scikit-learn>=1.4.0
python-multipart>=0.0.9
torch>=2.0.0
python-docx>=1.1.2
slowapi
python-dotenv
groq
